refactor: move GrailsResponseMutator to grails-web-common
